♻️ css 리팩토링
