+++ src/styles/commonStyles.jsx
@@ -40,49 +40,3 @@ export const Container = styled.div`
   position: relative;
   overflow: auto;
 `;
-
-// 검색창 스타일
-import { SearchOutlined } from '@ant-design/icons';
-import Link from 'next/link';
-export const SearchBarContainer = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: ${(props) => (props.header ? '460px' : '580px')};
-  height: ${(props) => (props.header ? '48px' : '64px')};
-  left: ${(props) => (props.header ? '0px' : 'auto')};
-  top: ${(props) => (props.header ? '0px' : 'auto')};
-  border: 2px solid var(--secondary);
-  border-radius: 50px;
-  padding-right: 20px;
-`;
-
-export const SearchInput = styled.input`
-  width: 100%;
-  height: 100%;
-  padding: 30px;
-  font-size: ${(props) => (props.header ? '14px' : '18px')};
-  font-weight: 400;
-  color: #666666;
-  background-color: transparent;
-  border: none;
-  outline: none;
-`;
-
-export const Icon = styled.div`
-  width: 32px;
-  height: 32px;
-  cursor: pointer;
-`;
-
-export const StyledSearchOutlined = styled(SearchOutlined)`
-  width: 32px;
-  height: 32px;
-  font-size: 30px;
-  color: #666666;
-`;
-
-export const ResetLink = styled(Link)`
-  text-decoration: none;
-  cursor: pointer;
-`;
